Better screen elements size unit

+++ app/src/main/java/com/any/some/data/room/entity/WhiteboardItemDataEntity.kt
@@ -11,6 +11,6 @@ class WhiteboardItemDataEntity(
     val body: String,
     val x: Int,
     val y: Int,
-    val width: Int,
-    val height: Int
+    val width: Float,
+    val height: Float
 )
